refactor: move ecstore erasure owner modules
