+++ crates/ecstore/src/lib.rs
@@ -23,8 +23,7 @@ mod data_movement;
 mod data_usage;
 mod diagnostics;
 mod disk;
-mod erasure_codec;
-mod erasure_coding;
+mod erasure;
 mod error;
 mod io_support;
 pub(crate) mod layout;
